ui: view updates
