+++ BDKSwiftExampleWallet/View Model/AmountViewModel.swift
@@ -1,0 +1,31 @@
+//
+//  AmountViewModel.swift
+//  BDKSwiftExampleWallet
+//
+//  Created by Matthew Ramsden on 9/22/23.
+//
+
+import Foundation
+
+@MainActor
+@Observable
+class AmountViewModel {
+    let bdkClient: BDKClient
+    var balanceTotal: UInt64?
+
+    init(bdkClient: BDKClient = .live) {
+        self.bdkClient = bdkClient
+    }
+
+    func getBalance() {
+        do {
+            let balance = try bdkClient.getBalance()
+            self.balanceTotal = balance.total
+        } catch let error as WalletError {
+            print("getBalance - Send Error: \(error.localizedDescription)")
+        } catch {
+            print("getBalance - Undefined Error: \(error.localizedDescription)")
+        }
+    }
+
+}
