+++ BDKSwiftExampleWallet/View/FeeView.swift
@@ -7,79 +7,7 @@
 
 import BitcoinDevKit
 import BitcoinUI
-import Observation
 import SwiftUI
-
-@MainActor
-@Observable
-class FeeViewModel {
-    let feeClient: FeeClient
-    let bdkClient: BDKClient
-
-    var txBuilderResult: TxBuilderResult?
-    var recommendedFees: RecommendedFees?
-    var selectedFeeIndex: Int = 2
-    var selectedFee: Int? {
-        guard let fees = recommendedFees else {
-            return nil
-        }
-        switch selectedFeeIndex {
-        case 0: return fees.minimumFee
-        case 1: return fees.hourFee
-        case 2: return fees.halfHourFee
-        default: return fees.fastestFee
-        }
-    }
-    var selectedFeeDescription: String {
-        guard let selectedFee = selectedFee else {
-            return "Failed to load fees"
-        }
-
-        let feeText = text(for: selectedFeeIndex)
-        return "Selected \(feeText) Fee: \(selectedFee) sats"
-    }
-    func text(for index: Int) -> String {
-
-        switch index {
-
-        //"Minimum Fee"
-        case 0:
-            return "No Priority"
-
-        //"Hour Fee"
-        case 1:
-            return "Low Priority"
-
-        //"Half Hour Fee"
-        case 2:
-            return "Medium Priority"
-
-        //"Fastest Fee"
-        case 3:
-            return "High Priority"
-
-        default:
-            return ""
-
-        }
-
-    }
-
-    init(feeClient: FeeClient = .live, bdkClient: BDKClient = .live) {
-        self.feeClient = feeClient
-        self.bdkClient = bdkClient
-    }
-
-    func getFees() async {
-        do {
-            let recommendedFees = try await feeClient.fetchFees()
-            self.recommendedFees = recommendedFees
-        } catch {
-            print("getFees error: \(error.localizedDescription)")
-        }
-    }
-
-}
 
 struct FeeView: View {
     let amount: String
